#1. Bug fix: moving floors stopped with line special #89 would cause NPE because function was set to null not NOP. However setting to NOP would cause the thinker to be removed from the list so the plat could never be restarted. So split NOP into HALT (which cannot be restarted) and WAIT (which can.)
This can be seen in E2M2 by circling the square pillar in sector 43,
starting & stopping the three circular lifts to the SouthEast.

#2. If a locking key (caps lock/num lock/scroll lock) is assigned to a
movement command then its toggle state is used (i.e. on if its light on
the keyboard is on) instead of its physical position. Likely to be useful
useful only for run & strafe.  Chocolate Doom has a very similar function.